+++ src/bin/uairctl/main.rs
@@ -1,63 +1,52 @@
-use std::env;
 use std::os::unix::net::UnixStream;
 use std::io::Write;
-use uair::Command;
-
-argwerk::define! {
-	/// Command-line interface for controlling uair.
-	#[derive(Default)]
-	#[usage = "uairctl [options..]"]
-	struct Args {
-		pause: bool,
-		resume: bool,
-		socket_path: String = if let Ok(xdg_runtime_dir) = env::var("XDG_RUNTIME_DIR") {
-			xdg_runtime_dir + "/uair.sock"
-		} else if let Ok(tmp_dir) = env::var("TMPDIR") {
-			tmp_dir + "/uair.sock"
-		} else {
-			"/tmp/uair.sock".into()
-		},
-		help: bool,
-	}
-	/// Pause the timer.
-	["-p" | "--pause"] => {
-		pause = true;
-	}
-	/// Resume the timer.
-	["-r" | "--resume"] => {
-		resume = true;
-	}
-	/// Specifies a socket file.
-	["-s" | "--socket", path] => {
-		socket_path = path;
-	}
-	/// Show help message and quit.
-	["-h" | "--help"] => {
-		help = true;
-	}
-}
+use gumdrop::Options;
+use serde::{Serialize, Deserialize};
+use uair::get_socket_path;
 
 fn main() -> anyhow::Result<()> {
-	let args = Args::args()?;
+	let mut args = Args::parse_args_default_or_exit();
 
-	if args.help {
-		println!("{}", Args::help());
-		return Ok(())
+	match (args.pause, args.resume) {
+		(true, true) => args.command = Some(Command::Toggle(NoArgs {})),
+		(true, false) => args.command = Some(Command::Pause(NoArgs {})),
+		(false, true) => args.command = Some(Command::Resume(NoArgs {})),
+		(false, false) => {}
 	};
+	let comm = if let Some(comm) = args.command { comm } else { return Ok(()) };
+	if args.socket.is_empty() { args.socket = get_socket_path() };
 
-	let comm = match (args.pause, args.resume) {
-		(true, true) => Command::Toggle,
-		(true, false) => Command::Pause,
-		(false, true) => Command::Resume,
-		(false, false) => {
-			println!("{}", Args::help());
-			return Ok(());
-		}
-	};
 	let command = bincode::serialize(&comm)?;
 
-	let mut stream = UnixStream::connect(&args.socket_path)?;
+	let mut stream = UnixStream::connect(&args.socket)?;
 	stream.write_all(&command)?;
 
 	Ok(())
 }
+
+#[derive(Options)]
+struct Args {
+	#[options(help = "Show help message and quit.")]
+	help: bool,
+	#[options(help = "Pause the timer. (Deprecated)")]
+	pause: bool,
+	#[options(help = "Resume the timer. (Deprecated)")]
+	resume: bool,
+	#[options(help = "Specifies the socket file.")]
+	socket: String,
+	#[options(command)]
+	command: Option<Command>,
+}
+
+#[derive(Options, Serialize, Deserialize)]
+enum Command {
+	#[options(help = "Pause the timer.")]
+	Pause(NoArgs),
+	#[options(help = "Resume the timer.")]
+	Resume(NoArgs),
+	#[options(help = "Toggle the state of the timer.")]
+	Toggle(NoArgs),
+}
+
+#[derive(Options, Serialize, Deserialize)]
+struct NoArgs {}